Changed int to unsigned long

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,15 @@
 int main(void)
 {
 	int i;
-	int t1 = 1, t2 = 2;
+	unsigned long t1 = 1, t2 = 2;
 
-	int nextTerm = t1 + t2;
+	unsigned long nextTerm = t1 + t2;
 
-	printf("%d, %d, ", t1, t2);
+	printf("%lu, %lu, ", t1, t2);
 
 	for (i = 3; i <= 100; ++i)
 	{
-		printf("%d, ", nextTerm);
+		printf("%lu, ", nextTerm);
 		t1 = t2;
 		t2 = nextTerm;
 		nextTerm = t1 + t2;
